Add check to see if backup is already in progress.

+++ src/main/java/cloud/stonehouse/s3backup/Backup.java
@@ -14,6 +14,7 @@ import java.util.TreeMap;
 
 class Backup extends BukkitRunnable {
 
+    private final Boolean backupInProgress;
     private final String customPrefix;
     private final S3Backup s3Backup;
     private final Player player;
@@ -22,55 +23,68 @@ class Backup extends BukkitRunnable {
         this.customPrefix = customPrefix;
         this.s3Backup = s3Backup;
         this.player = player;
-        s3Backup.sendMessage(player, "Saving worlds. This will take a moment");
-        Bukkit.dispatchCommand(Bukkit.getConsoleSender(), "save-off");
-        Bukkit.dispatchCommand(Bukkit.getConsoleSender(), "save-all");
-        s3Backup.sendMessage(player, "Worlds saved");
+
+        this.backupInProgress = s3Backup.inProgress();
+
+        if (!backupInProgress) {
+            s3Backup.sendMessage(player, "Saving worlds. This will take a moment");
+            Bukkit.getScheduler().callSyncMethod(s3Backup, () -> Bukkit.dispatchCommand(Bukkit.getConsoleSender(), "save-off"));
+            Bukkit.getScheduler().callSyncMethod(s3Backup, () -> Bukkit.dispatchCommand(Bukkit.getConsoleSender(), "save-all"));
+            s3Backup.sendMessage(player, "Worlds saved");
+        }
     }
 
     @Override
     public void run() {
-        try {
-            String dateFormat = s3Backup.getFileConfig().getBackupDateFormat();
-            if (s3Backup.illegalString(dateFormat)) {
-                throw new StringFormatException("Invalid date format. Only alphanumeric characters, underscores and " +
-                        "hyphens are permitted. Please check config.yml");
-            }
-            if (s3Backup.illegalPrefix(s3Backup.getFileConfig().getPrefix())) {
-                throw new StringFormatException("Invalid prefix format. Only alphanumeric characters, underscores, " +
-                        "hyphens and forward slashes are permitted. Please check config.yml");
-            }
+        if (backupInProgress) {
+            s3Backup.sendMessage(player, "A backup is already in progress");
 
-            String archiveName = customPrefix + new SimpleDateFormat(s3Backup.getFileConfig().getBackupDateFormat())
-                    .format(new Date()) + ".zip";
-            String backupDir = s3Backup.getFileConfig().getBackupDir();
-
-            s3Backup.sendMessage(player, "Backup initiated");
-            String archivePath = backupDir + File.separator + archiveName;
-
-            s3Backup.archive().zipFile(player, Paths.get("").toAbsolutePath().normalize().toString(), archivePath);
-            s3Backup.s3Put().put(archiveName);
-            s3Backup.archive().deleteFile(archivePath);
-
-            s3Backup.sendMessage(player, "Backup complete");
-
-            int maxBackups = s3Backup.getFileConfig().getMaxBackups();
-            if (maxBackups > 0) {
-                TreeMap<Date, S3ObjectSummary> backups = new S3List(s3Backup).list(null, false, 0);
-
-                while (backups.size() > maxBackups) {
-                    Date remove = backups.firstKey();
-
-                    s3Backup.s3Delete().delete(player, backups.get(remove).getKey().replace(s3Backup.getFileConfig()
-                            .getPrefix(), ""));
-                    backups.remove(remove);
+        } else {
+            s3Backup.setProgress(true);
+            try {
+                String dateFormat = s3Backup.getFileConfig().getBackupDateFormat();
+                if (s3Backup.illegalString(dateFormat)) {
+                    throw new StringFormatException("Invalid date format. Only alphanumeric characters, underscores and " +
+                            "hyphens are permitted. Please check config.yml");
                 }
-            }
+                if (s3Backup.illegalPrefix(s3Backup.getFileConfig().getPrefix())) {
+                    throw new StringFormatException("Invalid prefix format. Only alphanumeric characters, underscores, " +
+                            "hyphens and forward slashes are permitted. Please check config.yml");
+                }
 
-        } catch (Exception e) {
-            s3Backup.exception(player, "Backup failed", e);
+                String archiveName = customPrefix + new SimpleDateFormat(s3Backup.getFileConfig().getBackupDateFormat())
+                        .format(new Date()) + ".zip";
+                String backupDir = s3Backup.getFileConfig().getBackupDir();
+
+                s3Backup.sendMessage(player, "Backup initiated");
+                String archivePath = backupDir + File.separator + archiveName;
+
+                s3Backup.archive().zipFile(player, Paths.get("").toAbsolutePath().normalize().toString(), archivePath);
+                s3Backup.s3Put().put(archiveName);
+                s3Backup.archive().deleteFile(archivePath);
+
+                s3Backup.sendMessage(player, "Backup complete");
+
+                int maxBackups = s3Backup.getFileConfig().getMaxBackups();
+                if (maxBackups > 0) {
+                    TreeMap<Date, S3ObjectSummary> backups = new S3List(s3Backup).list(null, false, 0);
+
+                    while (backups.size() > maxBackups) {
+                        Date remove = backups.firstKey();
+
+                        s3Backup.s3Delete().delete(player, backups.get(remove).getKey().replace(s3Backup.getFileConfig()
+                                .getPrefix(), ""));
+                        backups.remove(remove);
+                    }
+                }
+
+            } catch (Exception e) {
+                s3Backup.exception(player, "Backup failed", e);
+            } finally {
+                Bukkit.getScheduler().callSyncMethod(s3Backup, () -> Bukkit.dispatchCommand(Bukkit.getConsoleSender(), "save-on"));
+                s3Backup.setProgress(false);
+            }
         }
 
-        Bukkit.getScheduler().callSyncMethod(s3Backup, () -> Bukkit.dispatchCommand(Bukkit.getConsoleSender(), "save-on"));
     }
 }
